Calendar_MyApp/requirements.txt
Flask
pandas
pdfplumber
openpyxl
python-docx
ics
werkzeug
